pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-quick"
version = "0.1.0"
description = "Lightning-fast Git workflows with AI-powered commit messages"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["git", "cli", "productivity", "ai", "commit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "gitpython>=3.1",
    "toml>=0.10",
    "openai>=1.0",
    "anthropic>=0.25",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
git-quick = "git_quick.cli:quick"
git-story = "git_quick.cli:story"
git-time = "git_quick.cli:time_track"
git-sync-all = "git_quick.cli:sync_all"

[project.urls]
Homepage = "https://github.com/yourusername/git-quick"
Documentation = "https://github.com/yourusername/git-quick#readme"
Repository = "https://github.com/yourusername/git-quick"

[tool.setuptools.packages.find]
where = ["."]
include = ["git_quick*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
